WL-712 Added restriction of the submission type when using Turnitin.
I have added mutual exclusion restrictions to the user interface for creating
new assignments. Specifically, I have enforced the restriction that the
submission type must be 'attachments only' when content-review (Turnitin) is
being used. This means that if you select to use content-review, 'attachments'
is automatically selected and likewise content-review is deselected if any
submission type other than attachments is specified. In each case, focus is then
given to the control which was changed in order to give a cue to the user that
something changed in the interface.