WL-712 Made the submission types & turnitin controls adjacent for clarity.
I moved the section of Turnitin controls to be underneath those for selecting
the student submission type. In conjunction with the previous commit (which
includes in some cases switching focus on the controls) this should make things
clearer to the user. I have also added an instructional note saying
that the submission type must be 'attachments only' when using Turnitin.