WL-903 Added restriction of the submission type when using Turnitin.
I have added mutual exclusion restrictions to the user interface for creating
new assignments. Specifically, I have enforced the restriction that the
submission type must be 'attachments only' when content-review (Turnitin) is
being used. This means that if you select to use content-review, 'attachments'
is automatically selected and likewise content-review is deselected if any
submission type other than attachments is specified. In each case, focus is then
given to the control which was changed in order to give a cue to the user that
something changed in the interface. (Note: this commit is analogous to one I
made for WL-712 on the 2-5-x branch, but I re-crafted this by hand as the
previous attempt was broken).

WL-938 Restricted to submission of single files when Turnitin is used.

This is very similar to a commit made against WL-679, however crucially that
commit involved applying the logic in the wrong place, so it caused other
related functionality to break (see WL-947, WL-868).

WL-925 Submission via a URL disabled when Turnitin is being used.

This is very similar to a commit made against WL-680, however crucially that
commit involved applying the logic in the wrong place, so it caused other
related functionality to break (see WL-947, WL-868).

WL-1013 Undone most of previous implementation in order to replace with another.

[*] Restored Turnitin controls to original location.
[*] Removed Javascript associated with enforcing 'attachments only' in UI.
[*] Message about Turnitin == 'attachments only' *remains*.

WL-1363 Error when noting for TII to have.
If none of the attachments associated with an assignment submission are accepted by TII then a NPE would have been generated.

WL-1913 Only allow single upload with Turnitin.